requirements.txt
Pyyaml==6.0.1﻿
pandas ==2.0.3
loguru==0.7.3
colorama==0.4.6
et-xmlfile==1.1.0
numpy==1.26.1
openpyxl==3.1.2
pyarrow==14.0.0
python-dateutil==2.8.2
pytz==2023.3
regex==2024.5.15
setuptools==80.9.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
Unidecode==1.3.6
win32-setctime==1.1.0
